requirements.txt
pyautogui
pytesseract
pillow
pywin32
autoit
python-dotenv
